+++ test/widgets/journal/tags/tag_add_test.dart
@@ -45,6 +45,7 @@ void main() {
         (_) => EntryState.dirty(
           entryId: testTextEntry.meta.id,
           entry: testTextEntry,
+          showMap: false,
         ),
       );